+++ state_beacon/lib/src/extensions/readable.dart
@@ -2,6 +2,7 @@ part of 'extensions.dart';
 
 extension ReadableBeaconUtils<T> on ReadableBeacon<T> {
   /// Converts a [ReadableBeacon] to [Stream]
+  /// The stream can only be canceled by calling [dispose]
   Stream<T> toStream({
     FutureOr<void> Function()? onCancel,
   }) {
@@ -17,10 +18,8 @@ extension ReadableBeaconUtils<T> on ReadableBeacon<T> {
       onCancel?.call();
     }
 
-    controller.onCancel = cancel;
-
     onDispose(cancel);
 
-    return controller.stream;
+    return controller.stream.asBroadcastStream();
   }
 }